efivar: Use struct guid_table instead of uuid_table
uuid_table was the linux name. When libefivar was started, I tried to
make it compatible with Linux. However, that's no longer relevant: (a)
little to no Linux code was subsequently ported and (b) Linux compat has
eroded. This erodes it a bit more to cope with the uuid_t -> efi_guid_t
changes. This also moves a couple of functions around to reduce copying and
updates consumers for the visible parts of this change.

Sponsored by:		Netflix
Reviewed by:		tsoome, kib
Differential Revision:	https://reviews.freebsd.org/D50060